Move fonts package to `assets/src`
