+++ assets/src/fonts/README.md
@@ -1,0 +1,7 @@
+# Fonts
+
+Provides a list of system fonts and fonts available from Google Fonts, prepared for use in the Web Stories editor.
+
+The list, `font.json`, contains additional information like metrics for each font, and is updated regularly using an automated CI task.
+
+To update the list, run `GOOGLE_FONTS_API_KEY=yourapikey node index.js`.
