+++ assets/src/fonts/utils/buildFonts.js
